requirements.txt
# DEPENDENCIAS PYTHON - ESTILO Y MODA E-COMMERCE
# Compatible con Python 3.11+

# FRAMEWORK WEB
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# VALIDACION DE DATOS
pydantic==2.5.0
annotated-types==0.6.0
typing_extensions==4.8.0

# BASE DE DATOS
pymongo==4.6.0
motor==3.3.2
dnspython==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP Y NETWORKING
httpx==0.25.2
httpcore==1.0.2
h11==0.14.0
certifi==2023.11.17
idna==3.6

# UTILIDADES GENERALES
python-dotenv==1.0.0
python-multipart==0.0.6
click==8.1.7
colorama==0.4.6

# PROGRAMACION ASINCRONA
anyio==4.2.0
sniffio==1.3.0

# MACHINE LEARNING
numpy==1.24.4
scikit-learn==1.3.2
joblib==1.3.2
threadpoolctl==3.2.0

# CACHE Y REDIS
redis==5.0.1

# SEGURIDAD Y AUTENTICACION
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography>=42.0.0

# VALIDACION Y UTILIDADES
email-validator==2.1.0
requests==2.31.0
urllib3==2.1.0

# CLOUDINARY - OPTIMIZACION DE IMAGENES
cloudinary==1.36.0
Pillow==10.1.0